refactor(user-management): move byte-identical base_service to autobot_shared (#12647)
First step of #12645's user_management de-fork. The package is forked across
both backends with 25 shared-but-divergent files, so the risk is reconciling
semantic drift -- except base_service.py, which is byte-identical in both and
imports nothing backend-specific (uuid, dataclasses, sqlalchemy, and
autobot_shared.logging_manager). That makes it the one move with no drift to
reconcile.

Note the issue lists base_service as 6 diff lines; it is now 0, so this step is
smaller and safer than when #12647 was written.

Creates autobot_shared/user_management/ as the consolidation target, and leaves
a re-export shim in each backend so all 12 existing importers keep working --
the fork goes, not the callers. Same pattern as auth_rbac re-exporting
is_admin_role (#12948).

Tests assert both shims resolve to the SAME class objects rather than merely
equivalent ones: two same-named classes are not interchangeable, which is the
trap #12913 fixed for CircuitState and #12932 records for the memory graph.

Verified: 6 new tests pass, 1137 passed across autobot_shared, the three
user_management service modules still import, flake8 clean.
